--- frontend/src/components/Contact.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Contact() {
-  return (
-    <div>Contact</div>
-  )
-}
-
-export default Contact
