--- frontend/src/components/Home.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Home() {
-  return (
-    <div>Home</div>
-  )
-}
-
-export default Home
