--- frontend/src/components/Service.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Service() {
-  return (
-    <div>Service</div>
-  )
-}
-
-export default Service
