+++ frontend/src/routes/AppRoutes.jsx
@@ -3,27 +3,22 @@ import { Routes, Route } from "react-router-dom";
 import LoadingBar from "../components/LoadingBar";
 // import Profile from "../components/Profile";
 
-const Home = lazy(() => import("../components/Home"));
-const About = lazy(() => import("../components/About"));
-const Contact = lazy(() => import("../components/Contact"));
-const Service = lazy(() => import("../components/Service"));
 const Register = lazy(() => import("../components/Register"));
-
-
+const Dashboard = lazy(() => import("../components/Dashboard"));
+const UserManagement = lazy(() => import("../components/UserMangement"));
+const Table = lazy(() => import("../components/ProjectsTable"));
 
 const AppRoutes = () => {
- 
   return (
-    <Suspense fallback={<LoadingBar/>}>
+    <Suspense fallback={<LoadingBar />}>
       <Routes>
-        <Route exact path="/" element={<Home />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/contact" element={<Contact />} />
-        <Route path="/services" element={<Service />} />
+        <Route path="/" element={<Dashboard />} />
         <Route path="/register" element={<Register />} />
+        <Route path="/dashboard" element={<Dashboard />} />
+        <Route path="/user-list" element={<UserManagement />} />
+        <Route path="/table" element={<Table />} />
 
         {/* <Route path="/profile" element={<Profile />} /> */}
-
       </Routes>
     </Suspense>
   );
